Added untested pixel resolution to metadata store
I added a series of functions across the program to test for the correct
metadata file, find the X and Y resolutions based on that metadata file.

I also straight up took the key from the Z resolution line in the Scan
Image comment and turned it into the global variable.  This was applied
in metadata store.

None of this was tested.